Refactor algorithms to solve generic problem.
Both `sparse_direct` and `sparse_steepest` solve a sequence of generic
IRWLS problems with minimal assumptions about the classification task.
This allows us to reuse the same code for both the linear and nonlinear
cases, as well as intercept, by merely changing the design matrix.

The `trainMM` function automatically builds the correct design matrix
based on information contained in a `BinaryClassifier` object.